requirements.txt
# RAG Playbook - All Dependencies
# Install with: pip install -r requirements.txt

# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langchain-mongodb>=0.2.0
langchain-community>=0.3.0

# MongoDB
pymongo[srv]>=4.6.0

# PDF Processing
pypdf>=4.0.0

# Environment & Utilities
python-dotenv>=1.0.0
certifi>=2024.0.0

# Hybrid Search (Step 03)
rank_bm25>=0.2.2

# Graph RAG (Step 04)
neo4j>=5.0.0
